initial commit
Initializes the xtree repo and adds the SSPL license to prevent open
washing or stripmining IP from Lucenia.

This project may eventually end up ASLv2, but for now we keep it SSPL
for onprem only.

If you're looking at this and steal code for your cloud service you are
not only engaging in bad faith behavior, it will be enforced by lucenia.

Signed-off-by: Nicholas Walter Knize <nknize@apache.org>