Add initial experimental MMap impl and clean up some memory leaks
This commit addresses some initial memory leaks, but not all.
It also adds a separate, experimental mmap implementation that
keeps the data nodes in a separate (xdata) file from the internal
nodes (xtree). This is so that we can mmap traverse the tree w/o
access the actual data unless absolutely necessary. It also enables
us to compress the data separate from the internal tree, giving us
the most flexibility in minimizing storage requirements.

Signed-off-by: Nicholas Walter Knize <nknize@apache.org>